Move the argument parser to the datamatic.py file
