+++ datamatic.py
@@ -1,8 +1,38 @@
 """
 A tool for generating code based on a schema of components.
 """
+import argparse
+import pathlib
 from datamatic import main
 
+
+def parse_args():
+    """
+    Read the command line.
+    """
+    parser = argparse.ArgumentParser(
+        description="Given a component spec and a directory, scan the "
+                    "directory for dm and dmx files and generate the "
+                    "appropriate files"
+    )
+
+    parser.add_argument(
+        "-s", "--spec",
+        required=True,
+        type=pathlib.Path,
+        help="A path to the component spec JSON file"
+    )
+
+    parser.add_argument(
+        "-d", "--dir",
+        required=True,
+        type=pathlib.Path,
+        help="A path to the directory to scan for dm and dmx files"
+    )
+
+    return parser.parse_args()
+
+
 if __name__ == "__main__":
-    args = main.parse_args()
+    args = parse_args()
     main.main(args.spec, args.dir)
