+++ datamatic/main.py
@@ -28,33 +28,6 @@ def fill_flag_defaults(spec):
             attr["flags"] = {**defaults, **attr_flags}
 
 
-def parse_args():
-    """
-    Read the command line.
-    """
-    parser = argparse.ArgumentParser(
-        description="Given a component spec and a directory, scan the "
-                    "directory for dm and dmx files and generate the "
-                    "appropriate files"
-    )
-
-    parser.add_argument(
-        "-s", "--spec",
-        required=True,
-        type=pathlib.Path,
-        help="A path to the component spec JSON file"
-    )
-
-    parser.add_argument(
-        "-d", "--dir",
-        required=True,
-        type=pathlib.Path,
-        help="A path to the directory to scan for dm and dmx files"
-    )
-
-    return parser.parse_args()
-
-
 def main(specfile: pathlib.Path, directory: pathlib.Path):
     """
     Entry point.
